Renaming files, reverting to previous Game constructor.
